requirements.txt
opencv-python~=4.10.0.84
pillow~=10.4.0
python-dateutil~=2.9.0
Django~=4.2.17
djangorestframework~=3.15.2
celery~=5.4.0
django-celery-beat~=2.7.0
redis~=5.2.1
pytest-django~=4.9.0
open-meteo~=0.3.2
django-redis~=5.4.0
spotipy~=2.25.1
pre-commit~=4.3.0